refactor: change file strucutre to a service-like one
