+++ tests/test_rag.py
@@ -6,8 +6,8 @@ import pytest
 import asyncio
 import os
 from unittest.mock import patch, MagicMock, AsyncMock
-from rag.rag import RAG, SlidingWindowChunking, crawl4ai_crawl_many, CrawlResult, DB_Site
-from rag.search import Searcher
+from services.rag.rag import RAG, SlidingWindowChunking, crawl4ai_crawl_many, CrawlResult, DB_Site
+from services.rag.search import Searcher
 from crawl4ai import AsyncWebCrawler, CacheMode, BrowserConfig, CrawlerRunConfig, CrawlResult, MarkdownGenerationResult
 from crawl4ai.markdown_generation_strategy import DefaultMarkdownGenerator
 from crawl4ai.content_filter_strategy import PruningContentFilter
@@ -18,9 +18,9 @@ from pydantic import BaseModel
 
 @pytest.fixture
 def mock_rag():
-    with patch('rag.rag.Searcher') as MockSearcher, \
-         patch('rag.rag.VecDb') as MockVecDb, \
-         patch('rag.rag.RAGOpenai') as MockRAGOpenai:
+    with patch('services.rag.rag.Searcher') as MockSearcher, \
+         patch('services.rag.rag.VecDb') as MockVecDb, \
+         patch('services.rag.rag.RAGOpenai') as MockRAGOpenai:
         MockSearcher.return_value = AsyncMock()
         MockVecDb.return_value = AsyncMock()
         MockRAGOpenai.return_value = AsyncMock()
@@ -36,7 +36,7 @@ async def test_rag_initialization(mock_rag):
 @pytest.mark.asyncio
 async def test_search_and_crawl(mock_rag):
     mock_rag.sr.search.return_value = [SearchItem(title="Example", url="http://example.com", is_source_local=False, is_source_both=False)]
-    with patch('rag.rag.crawl4ai_crawl_many', new_callable=AsyncMock) as mock_crawl:
+    with patch('services.rag.rag.crawl4ai_crawl_many', new_callable=AsyncMock) as mock_crawl:
         mock_crawl.return_value = [CrawlResult(url="http://example.com", html="<html></html>", success=True)]
         results = await mock_rag.search_and_crawl("test query")
         assert len(results) == 1
